+++ src/lib/component.script.js
@@ -1,0 +1,3 @@
+import defineComponent from './component.js';
+
+(typeof gobalThis !== 'undefined' ? globalThis : window).defineComponent = defineComponent;