+++ tools/workbench/log.js
@@ -1,5 +1,0 @@
-module.exports = (...args) => {
-  if (!process.silence) {
-    console.log(new Date().toISOString(), ...args);
-  }
-}